+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-22T09:52:31.635Z",
+    "2026-04-22T11:07:47.246Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {